Missing docstring.

Missing docstring for wildboottest. Check Documenter's build log for details.

WildBootTests.AuxWtType — Type

Auxilliary weight types: rademacher, mammen, webb, normal, gamma

WildBootTests.PType — Type

p value types: symmetric, equaltail, lower, upper

WildBootTests.MAdjType — Type

Multiple hypothesis adjustment types: nomadj, bonferroni, sidak

WildBootTests.DistStatType — Type

Bootstrap distribution statistics optionally returned

WildBootTests.teststat — Function

Return test statistic subject to wild bootstrap test

WildBootTests.stattype — Function

Return type of test statistic subject to wild bootstrap test: "t", "z", "F", or "χ²"

WildBootTests.p — Function

Return p value from wild bootstrap test

WildBootTests.padj — Function

Returnp p value from wild bootstrap test after multiple-hypothesis adjustment, if any

WildBootTests.reps — Function

Return requested number of replications in wild bootstrap test

WildBootTests.repsfeas — Function

Return actual number of replications in wild bootstrap test, subject to enumeration of Rademacher draws

WildBootTests.NBootClust — Function

Return number of bootstrapping clusters in wild bootstrap test

WildBootTests.dof — Function

Return degrees of freedom wild bootstrap test

WildBootTests.dof_r — Function

Return residual degrees of freedom wild bootstrap test

WildBootTests.plotpoints — Function

Return data for confidence plot of wild bootstrap test. Return value is a 2-tuple with named entries X and p holding the confidence sampling locations and p values respectively. X is in turn a 1- or 2-tuple of vectors of sampling coordinates for each dimension of the tested hypothesis.

WildBootTests.peak — Function

Return parameter value with peak p value in wild bootstrap test

WildBootTests.CI — Function

Return confidence interval matrix from wild bootstrap test, one row per disjoint piece

WildBootTests.dist — Function

Return bootstrap distribution of statistic or statistic numerator in wild bootstrap test

WildBootTests.statnumer — Function

Return numerator of test statistic in wild bootstrap test

WildBootTests.statvar — Function

Return denominator of test statistic in wild bootstrap test

WildBootTests.auxweights — Function

Return auxilliary weight matrix for wild bootstrap